added nft.storage,modified contract and test script
