+++ frontend-contracts/artifacts/@openzeppelin/contracts/access/AccessControl.sol/AccessControl.dbg.json
@@ -1,4 +1,4 @@
 {
   "_format": "hh-sol-dbg-1",
-  "buildInfo": "../../../../build-info/433386aaf183d3b7bbd32f02137ad30f.json"
+  "buildInfo": "../../../../build-info/80fce95c1d93b378557f052f50d868a3.json"
 }
