+++ frontend-contracts/artifacts/contracts/LazyNFT.sol/LazyNFT.dbg.json
@@ -1,4 +1,4 @@
 {
   "_format": "hh-sol-dbg-1",
-  "buildInfo": "../../build-info/2a76e64911174f24e25d192e1f83cc0f.json"
+  "buildInfo": "../../build-info/137fd2d6f852d5ceef3bfa6a87a5db7b.json"
 }
