+++ frontend-contracts/contracts/LazyNFT.sol
@@ -2,14 +2,14 @@
 pragma solidity ^0.8.4;
 pragma experimental ABIEncoderV2;
 
-import "hardhat/console.sol";
-import "@openzeppelin/contracts/access/AccessControl.sol";
-import "@openzeppelin/contracts/token/ERC721/ERC721.sol";
-import "@openzeppelin/contracts/token/ERC721/extensions/ERC721URIStorage.sol";
-import "@openzeppelin/contracts/utils/cryptography/ECDSA.sol";
-import "@openzeppelin/contracts/utils/cryptography/draft-EIP712.sol";
-import "@openzeppelin/contracts/utils/Counters.sol";
-import "hardhat/console.sol";
+import "hardhat/console.sol"; 
+import "@openzeppelin/contracts/access/AccessControl.sol";  
+import "@openzeppelin/contracts/token/ERC721/ERC721.sol"; 
+import "@openzeppelin/contracts/token/ERC721/extensions/ERC721URIStorage.sol";  
+import "@openzeppelin/contracts/utils/cryptography/ECDSA.sol";  
+import "@openzeppelin/contracts/utils/cryptography/draft-EIP712.sol"; 
+import "@openzeppelin/contracts/utils/Counters.sol";  
+import "hardhat/console.sol"; 
 
 contract LazyNFT is ERC721URIStorage, EIP712, AccessControl {
   using ECDSA for bytes32;
@@ -52,15 +52,15 @@ contract LazyNFT is ERC721URIStorage, EIP712, AccessControl {
     string collection;
   }
 
-  /// @notice Redeems an NFTVoucher for an actual NFT, creating it in the process.
-  /// @param redeemer The address of the account which will receive the NFT upon success.
-  /// @param voucher An NFTVoucher that describes the NFT to be redeemed.
-  /// @param signature An EIP712 signature of the voucher, produced by the NFT creator.
-  function redeem(address redeemer, NFTVoucher calldata voucher, bytes memory signature) public payable returns (uint256) {
-    // make sure signature is valid and get the address of the signer
-    address signer = _verify(voucher, signature);
+  struct NFTCID {
+    string v_url;
+  }
 
-    console.log(signer);
+
+  function redeem(address redeemer, NFTVoucher calldata voucher, NFTCID calldata meta,bytes memory signature) public payable returns (uint256) {
+    // make sure signature is valid and get the address of the signer
+    address signer = _verify(meta, signature);
+
 
     // make sure that the signer is authorized to mint NFTs
     require(hasRole(MINTER_ROLE, signer), "Signature invalid or unauthorized");
@@ -109,29 +109,23 @@ contract LazyNFT is ERC721URIStorage, EIP712, AccessControl {
     return pendingWithdrawals[msg.sender];
   }
 
-  /// @notice Returns a hash of the given NFTVoucher, prepared using EIP712 typed data hashing rules.
-  /// @param voucher An NFTVoucher to hash.
-  function _hash(NFTVoucher calldata voucher) internal view returns (bytes32) {
+
+  
+  function _hash(NFTCID calldata meta) internal view returns (bytes32) {
 
     return _hashTypedDataV4(keccak256(abi.encode(
-      keccak256("NFTVoucher(uint256 tokenId,uint256 minPrice,string uri,string collection)"),
-      //voucher.tokenId,
-      voucher.tokenId,
-      voucher.minPrice,
-      keccak256(bytes(voucher.uri)),
-      keccak256(bytes(voucher.collection))
+      keccak256("NFTCID(string v_url)"),
+      keccak256(bytes(meta.v_url))
     )));
   }
 
-  /// @notice Verifies the signature for a given NFTVoucher, returning the address of the signer.
-  /// @dev Will revert if the signature is invalid. Does not verify that the signer is authorized to mint NFTs.
-  /// @param voucher An NFTVoucher describing an unminted NFT.
-  /// @param signature An EIP712 signature of the given voucher.
-  function _verify(NFTVoucher calldata voucher, bytes memory signature) internal view returns (address) {
-    bytes32 digest = _hash(voucher);
+
+  
+  function _verify(NFTCID calldata meta, bytes memory signature) internal view returns (address) {
+    bytes32 digest = _hash(meta);
     return digest.toEthSignedMessageHash().recover(signature);
   }
-
+  
   function supportsInterface(bytes4 interfaceId) public view virtual override (AccessControl, ERC721) returns (bool) {
     return ERC721.supportsInterface(interfaceId) || AccessControl.supportsInterface(interfaceId);
   }
