+++ frontend-contracts/hardhat.config.js
@@ -15,14 +15,7 @@ module.exports = {
     hardhat: {
       chainId: 1337,
     },
-    mumbai: {
-      url: `https://polygon-mumbai.g.alchemy.com/v2/${PROJECT_ID}`,
-      accounts: [PRIVATE_KEY],
-    },
-    mainnet: {
-      url: `https://polygon-mainnet.infura.io/v3/${PROJECT_ID}`,
-      accounts: [PRIVATE_KEY],
-    },
+    
   },
   solidity: "0.8.4",
 };
