+++ frontend-contracts/newconfig.js
@@ -1,5 +1,4 @@
 
-  export const nftaddress = "0x96661790622a98F72246178c545C19A675e1aB10"
-  export const redeemeraddress = "some shit not needed"
-  export const minteraddress = "0xA973c909f0B8f096dA010903A91e0Fd22f8739Fd"
+  export const nftmarketaddress = "0x5FbDB2315678afecb367f032d93F642f64180aa3"
+  export const nftaddress = "0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512"
