+++ frontend-contracts/package.json
@@ -16,25 +16,25 @@
     "@nomiclabs/hardhat-ethers": "^2.0.2",
     "@nomiclabs/hardhat-waffle": "^2.0.1",
     "@openzeppelin/contracts": "^4.1.0",
+    "autoprefixer": "^10.3.7",
     "axios": "^0.23.0",
     "chai": "4.3.4",
     "dotenv": "^10.0.0",
+    "eslint": "8.1.0",
+    "eslint-config-next": "11.1.2",
     "ethereum-waffle": "^3.4.0",
     "ethers": "5.3.1",
     "ethers-eip712": "^0.2.0",
     "hardhat": "^2.6.7",
     "ipfs-http-client": "50.1.2",
     "next": "11.1.2",
-    "nft.storage": "^5.0.0",
+    "nft.storage": "^5.2.3",
+    "postcss": "^8.3.11",
     "react": "17.0.2",
     "react-dom": "17.0.2",
+    "tailwindcss": "^2.2.17",
     "typescript": "^4.5.3",
-    "web3modal": "^1.9.4",
-    "autoprefixer": "^10.3.7",
-    "eslint": "8.1.0",
-    "eslint-config-next": "11.1.2",
-    "postcss": "^8.3.11",
-    "tailwindcss": "^2.2.17"
+    "web3modal": "^1.9.4"
   },
   "devDependencies": {}
 }
